16KB seems to give better performance than 64KB
Number of system calls is lower than on master, but performance is erratic.
I need a better system to test this on to determin if this is really a good change.`

Upon further analysis syscall count is roughly half compared to the master branch.
Total runtime is still erratic but is inline with master while sometimes being faster
and sometimes slower. The FS read ahead behavior is probably to blame/thank there, but
¯\_(ツ)_/¯

Given the reduction in syscalls I'm going to call this a win. In the event of a multidisk
system or a system with a more performant FS I expect this change to improve performance.